fix listuserActivity
